+++ .github/workflows/tests.yml
@@ -60,7 +60,7 @@ jobs:
 
       - name: Configure environment
         run: |
-          echo "JWKS_BASE64=$(python generate_jwk.py)" >> $GITHUB_ENV
+          echo "JWKS_BASE64=$(python manage.py generate_jwks)" >> $GITHUB_ENV
 
       - name: Run celery worker for tests
         run: celery -A config.celery_app worker -B --loglevel=INFO --detach
